Adds rpc models module and moves data structures from service module
