+++ identity/src/lib/rpc/mod.rs
@@ -1,3 +1,4 @@
+pub mod models;
 pub mod server;
 pub mod service;
