+++ identity/tests/rpc_endpoints.rs
@@ -8,7 +8,7 @@ use tarpc::context;
 
 use helpers::rpc::Error;
 use identity::db::schema::{users::dsl::users, users_roles::dsl::users_roles};
-use identity::rpc::service::{Role, SessionInfo, User, UserRole};
+use identity::rpc::models::{Role, SessionInfo, User, UserRole};
 use identity::rpc::{rpc_client, rpc_server, service::IdentityServiceClient};
 use identity::{config::Configuration, db::Db, session::jwt::Jwt};
